+++ src/AtomUI.Desktop.Controls/Result/Result.SemanticParts.cs
@@ -1,0 +1,35 @@
+using AtomUI.Theme;
+using Avalonia.Controls;
+using Avalonia.Controls.Presenters;
+
+namespace AtomUI.Desktop.Controls;
+
+[SemanticPart(
+    "icon",
+    SelectorClass = "semantic-icon",
+    ContractType = typeof(Control),
+    Cardinality = SemanticPartCardinality.Multiple,
+    Since = "6.0")]
+[SemanticPart(
+    "title",
+    SelectorClass = "semantic-title",
+    ContractType = typeof(ContentPresenter),
+    Since = "6.0")]
+[SemanticPart(
+    "subTitle",
+    SelectorClass = "semantic-sub-title",
+    ContractType = typeof(ContentPresenter),
+    Since = "6.0")]
+[SemanticPart(
+    "extra",
+    SelectorClass = "semantic-extra",
+    ContractType = typeof(ContentPresenter),
+    Since = "6.0")]
+[SemanticPart(
+    "body",
+    SelectorClass = "semantic-body",
+    ContractType = typeof(ContentPresenter),
+    Since = "6.0")]
+public partial class Result
+{
+}